requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.36.0
sentencepiece>=0.1.99
pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.66.0
pyyaml>=6.0
tensorboard>=2.14.0
